requirements.txt
groq==0.18.0
python-telegram-bot==20.8
python-dotenv==1.0.1
flask==3.0.0